// src/lib/http/generated/models/appointment-by-id-response-dto-output-animal.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Cuidados com Animais
 * OpenAPI spec version: 1.0
 */

import type { AppointmentByIdResponseDtoOutputAnimalAge } from "./appointment-by-id-response-dto-output-animal-age";
import type { AppointmentByIdResponseDtoOutputAnimalAsset } from "./appointment-by-id-response-dto-output-animal-asset";
import type { AppointmentByIdResponseDtoOutputAnimalBreed } from "./appointment-by-id-response-dto-output-animal-breed";

export type AppointmentByIdResponseDtoOutputAnimal = {
	id: string;
	name: string;
	asset?: AppointmentByIdResponseDtoOutputAnimalAsset;
	breed: AppointmentByIdResponseDtoOutputAnimalBreed;
	age?: AppointmentByIdResponseDtoOutputAnimalAge;
};