// src/lib/http/generated/models/appointment-by-id-response-dto-output-client.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Cuidados com Animais
 * OpenAPI spec version: 1.0
 */
import type { AppointmentByIdResponseDtoOutputClientAvatar } from "./appointment-by-id-response-dto-output-client-avatar";

export type AppointmentByIdResponseDtoOutputClient = {
	id: string;
	name: string;
	email: string;
	avatar?: AppointmentByIdResponseDtoOutputClientAvatar;
};