// src/lib/http/generated/models/appointment-by-id-response-dto-output-coat-type.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Cuidados com Animais
 * OpenAPI spec version: 1.0
 */

export type AppointmentByIdResponseDtoOutputCoatType =
	(typeof AppointmentByIdResponseDtoOutputCoatType)[keyof typeof AppointmentByIdResponseDtoOutputCoatType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AppointmentByIdResponseDtoOutputCoatType = {
	short: "short",
	medium: "medium",
	long: "long",
	curly: "curly",
} as const;